requirements.txt
Django==3.1.7
django-cors-headers==3.7.0
djangorestframework==3.12.3
djangorestframework-simplejwt==4.6.0
gunicorn==20.1.0
python-dotenv==0.16.0
whitenoise==5.2.0
